Add a FLOSS variant
Moves all the Google Analytics code into the google variant and adds a
floss variant that eventually avoids having any non-free libraries.
