+++ app/src/floss/java/fi/bitrite/android/ws/WSAndroidApplication.java
@@ -1,0 +1,9 @@
+package fi.bitrite.android.ws;
+
+import fi.bitrite.android.ws.di.AppInjector;
+
+public class WSAndroidApplication extends BaseWSAndroidApplication {
+    protected AppInjector inject() {
+        return AppInjector.create(this);
+    }
+}
