+++ app/src/floss/java/fi/bitrite/android/ws/util/Tools.java
@@ -1,0 +1,14 @@
+package fi.bitrite.android.ws.util;
+
+import android.content.Context;
+
+/**
+ * General simple tools, mostly public methods.
+ */
+public class Tools extends BaseTools {
+
+    // Send a report to Google Analytics about  category/action
+    static public void gaReportException(Context context, String category, String action) {
+        // This just does nothing.
+    }
+}
