+++ app/src/floss/res/values-es/strings.xml
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_version">Versión %1$s (FLOSS)</string>
+</resources>
