+++ app/src/floss/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_version">Version %1$s (FLOSS)</string>
+</resources>
