+++ app/src/google/res/values/integers.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <bool name="prefs_ga_collect_stats_default">true</bool>
+</resources>
