+++ app/src/main/res/values-de/strings.xml
@@ -112,10 +112,6 @@
     <string name="prefs_message_refresh_interval_min_dialog_message">Wähle aus, wie häufig deine Mitteilungen automatisch vom Server geladen werden sollen.</string>
     <string name="prefs_message_refresh_interval_min_hint">Intervall (in Minuten)</string>
 
-    <string name="prefs_ga_collect_stats_title">Anonyme Benutzungsstatistiken senden</string>
-    <string name="prefs_ga_collect_stats_summary_on">Nutzungsstatistiken werden gelegentlich an Google Analytics gesendet, um den Entwickelnden die Verbesserung der App zu erleichtern</string>
-    <string name="prefs_ga_collect_stats_summary_off">Gerade werden keine Statistiken zur Verbesserung der App an Google Analytics versandt</string>
-
     <string name="prefscat_dev_title">Entwicklereinstellungen</string>
     <string name="prefs_dev_simulate_no_network_title">Simuliere Kein Netz</string>
