+++ app/src/main/res/values-es/strings.xml
@@ -99,10 +99,6 @@
     <string name="distance_unit_kilometers_short">km</string>
     <string name="distance_unit_miles_short">mi</string>
 
-    <string name="prefs_ga_collect_stats_title">Enviar estadísticas anónimas</string>
-    <string name="prefs_ga_collect_stats_summary_on">Enviar estadísticas a Google Analytics ayuda a los desarrolladores a entender el uso de la aplicación y a mejorarla, pero puedes desinscribirte.</string>
-    <string name="prefs_ga_collect_stats_summary_off">Enviar estadísticas a Google Analytics ayuda a los desarrolladores a entender el uso de la aplicación y a mejorarla, pero puedes desinscribirte.</string>
-
     <!-- Non-english speakers might prefer the default location to be over Africa, Europe & Central Asia. -->
     <!-- Sorry to all Brits (who default to the US) and rest of the non-western world. -->
     <string name="prefs_map_location_latitude_default">40.0000</string>
