+++ app/src/main/res/values-fr/strings.xml
@@ -99,10 +99,6 @@
     <string name="distance_unit_kilometers_short">km</string>
     <string name="distance_unit_miles_short">mi</string>
 
-    <string name="prefs_ga_collect_stats_summary_on">Envoyer des données statistiques à Google Analytics aide les développeurs à comprendre l\'usage de l\'application et à l\'améliorer, mais vous pouvez désactiver cette option.</string>
-    <string name="prefs_ga_collect_stats_summary_off">Envoyer des données statistiques à Google Analytics aide les développeurs à comprendre l\'usage de l\'application et à l\'améliorer, mais vous pouvez désactiver cette option.</string>
-    <string name="prefs_ga_collect_stats_title">Envoyer des données anonymes</string>
-
     <!-- Non-english speakers might prefer the default location to be over Africa, Europe & Central Asia. -->
     <!-- Sorry to all Brits (who default to the US) and rest of the non-western world. -->
     <string name="prefs_map_location_latitude_default">40.0000</string>
