+++ app/src/main/res/values/integers.xml
@@ -23,6 +23,5 @@
     <bool name="simulate_network_disconnected">false</bool>
 
     <integer name="prefs_message_refresh_interval_min_default">5</integer>
-    <bool name="prefs_ga_collect_stats_default">true</bool>
     <bool name="prefs_dev_simulate_no_network_default">false</bool>
 </resources>
