+++ app/src/main/res/values/strings.xml
@@ -118,12 +118,6 @@
     <string name="prefs_message_refresh_interval_min_dialog_message">Specify how often your messages should be refreshed from the server.</string>
     <string name="prefs_message_refresh_interval_min_hint">Interval (minutes)</string>
 
-    <string name="ga_property_id" translatable="false">UA-13228014-2</string>
-    <string name="prefs_ga_collect_stats_key" translatable="false">ga_collect_stats</string>
-    <string name="prefs_ga_collect_stats_title">Send anonymous statistics</string>
-    <string name="prefs_ga_collect_stats_summary_on">Sending statistics to Google Analytics helps the developers to understand usage and improve the app, but you can opt out</string>
-    <string name="prefs_ga_collect_stats_summary_off">Currently, your usage is not tracked in order to be sent to Google Analytics to help developers in their improvement endeavours</string>
-
     <string name="prefscat_dev_title">Developer settings</string>
     <string name="prefs_dev_simulate_no_network_key" translatable="false">dev_simulate_no_network</string>
     <string name="prefs_dev_simulate_no_network_title">Simulate No Service</string>
